Minimally functional shell feature complete
