+++ cmd/streams/stream.go
@@ -1,0 +1,37 @@
+package streams
+
+import (
+	"github.com/moby/term"
+)
+
+// The following code is originally from the github.com/docker/cli/cli/streams repo that
+// was "mindfully" (blindly) copy-pasted here due to Go module/dependency issues that prevented reuse
+
+// commonStream is an input stream used by the DockerCli to read user input
+type commonStream struct {
+	fd         uintptr
+	isTerminal bool
+	state      *term.State
+}
+
+// FD returns the file descriptor number for this stream
+func (s *commonStream) FD() uintptr {
+	return s.fd
+}
+
+// IsTerminal returns true if this stream is connected to a terminal
+func (s *commonStream) IsTerminal() bool {
+	return s.isTerminal
+}
+
+// RestoreTerminal restores normal mode to the terminal
+func (s *commonStream) RestoreTerminal() {
+	if s.state != nil {
+		term.RestoreTerminal(s.fd, s.state)
+	}
+}
+
+// SetIsTerminal sets the boolean used for isTerminal
+func (s *commonStream) SetIsTerminal(isTerminal bool) {
+	s.isTerminal = isTerminal
+}
